Continuing on D3D12 and RHI improvements

+++ Volt/Volt/src/Volt/Rendering/RenderGraph/RenderGraphCommon.h
@@ -6,6 +6,6 @@ namespace Volt
 {
 	namespace RenderGraphCommon
 	{
-		inline static constexpr uint32_t MAX_PASS_CONSTANTS_SIZE = 256;
+		inline static constexpr uint64_t MAX_PASS_CONSTANTS_SIZE = 256;
 	}
 }
